refactor: dto 레이어 서브패키지 분리 (account/connection/summary/sync/manual)
