+++ src/main/java/com/team/independence/asset/dto/account/AssetAccountDetailItem.java
@@ -1,4 +1,4 @@
-package com.team.independence.asset.dto;
+package com.team.independence.asset.dto.account;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
