+++ src/main/java/com/team/independence/asset/dto/account/LoanAccountQueryItem.java
@@ -1,19 +1,16 @@
-package com.team.independence.asset.dto;
+package com.team.independence.asset.dto.account;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
-import java.time.LocalDate;
-
 @Getter
 @Setter
 @NoArgsConstructor
-public class LoanAccountDetailItem {
+public class LoanAccountQueryItem {
+    private Long id;
     private String institutionName;
     private String loanName;
     private String accountDisplay;
     private Long loanBalance;
-    private LocalDate startDate;
-    private LocalDate endDate;
 }
