+++ src/main/java/com/team/independence/asset/dto/connection/UnlinkOrganizationResponse.java
@@ -1,4 +1,4 @@
-package com.team.independence.asset.dto;
+package com.team.independence.asset.dto.connection;
 
 import lombok.Builder;
 import lombok.Getter;
