+++ src/main/java/com/team/independence/asset/dto/manual/ManualAssetRequest.java
@@ -1,4 +1,4 @@
-package com.team.independence.asset.dto;
+package com.team.independence.asset.dto.manual;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
