+++ src/main/java/com/team/independence/asset/dto/manual/ManualAssetResponse.java
@@ -1,4 +1,4 @@
-package com.team.independence.asset.dto;
+package com.team.independence.asset.dto.manual;
 
 import lombok.Builder;
 import lombok.Getter;
